+++ src/winstan.ts
@@ -1,0 +1,6 @@
+export { log } from './logger';
+export { setColors } from './conf/color';
+export { setLevel } from './conf/level';
+export { setLocale } from './conf/locale';
+export { setService } from './conf/service';
+export { setTimeZone } from './conf/tz';